+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.12.8 (unreleased)
 
 - Add support for discovery via SSDP
+- Make IGD aware that certain actions live on WANPPP or WANIPC service
 
 
 0.12.7 (2018-10-18)